services/ai-decision/requirements.txt
# Core Dependencies
boto3>=1.34.0
psycopg2-binary>=2.9.9
redis>=5.0.1

# OpenSearch and Authentication
opensearch-py>=2.4.2
requests-aws4auth>=1.2.3

# HTTP Client
requests>=2.31.0

# Data Processing
python-dateutil>=2.8.2

# JSON Repair (for fixing malformed AI responses)
json-repair>=0.30.0

# Environment Management (for local development)
python-dotenv>=1.0.0

# Testing
pytest>=7.4.3
pytest-mock>=3.12.0
pytest-cov>=4.1.0

# Type Hints (optional but recommended)
typing-extensions>=4.9.0